Move registration under onboarding
