+++ packages/mobile/src/onboarding/registration/RegulatoryTerms.tsx
